requirements.txt
streamlit
opencv-python-headless
numpy
Pillow
matplotlib
streamlit-drawable-canvas
